Added Delete Service method for R2

+++ internal/service/r2.go
@@ -3,8 +3,9 @@ package service
 import (
 	"context"
 	"fmt"
+	"io"
 	"mime/multipart"
-	"path/filepath"
+	"strings"
 	"time"
 
 	"github.com/aws/aws-sdk-go-v2/aws"
@@ -40,7 +41,9 @@ func (s *R2Service) UploadFile(file *multipart.FileHeader, folder string) (strin
 
 	// Benzersiz dosya adı oluştur
 	filename := fmt.Sprintf("%d_%s", time.Now().Unix(), file.Filename)
-	key := filepath.Join(folder, filename)
+	key := strings.Join([]string{folder, filename}, "/")
+
+	fmt.Printf("R2 - Dosya yükleme işlemi başlatıldı. Key: %s\n", key)
 
 	// Upload işlemi
 	_, err = s.client.PutObject(context.TODO(), &s3.PutObjectInput{
@@ -51,19 +54,36 @@ func (s *R2Service) UploadFile(file *multipart.FileHeader, folder string) (strin
 	})
 
 	if err != nil {
+		fmt.Printf("R2 - HATA: Dosya yüklenirken hata oluştu. Key: %s, Hata: %v\n", key, err)
 		return "", err
 	}
 
-	// Public URL döndür
-	return fmt.Sprintf("https://%s.r2.cloudflarestorage.com/%s", s.bucketName, key), nil
+	publicURL := fmt.Sprintf("https://%s.r2.cloudflarestorage.com/%s", s.bucketName, key)
+	fmt.Printf("R2 - Dosya başarıyla yüklendi. Public URL: %s\n", publicURL)
+	return publicURL, nil
 }
 
 func (s *R2Service) DeleteFile(key string) error {
+	fmt.Printf("R2 - Dosya silme işlemi başlatıldı. Key: %s\n", key)
+
+	if key == "" {
+		fmt.Println("R2 - HATA: Boş key ile silme işlemi yapılamaz")
+		return fmt.Errorf("dosya key'i boş olamaz")
+	}
+
+	fmt.Printf("R2 - DeleteObject çağrısı yapılıyor. Bucket: %s, Key: %s\n", s.bucketName, key)
 	_, err := s.client.DeleteObject(context.TODO(), &s3.DeleteObjectInput{
 		Bucket: aws.String(s.bucketName),
 		Key:    aws.String(key),
 	})
-	return err
+
+	if err != nil {
+		fmt.Printf("R2 - HATA: Dosya silinirken hata oluştu. Key: %s, Hata: %v\n", key, err)
+		return err
+	}
+
+	fmt.Printf("R2 - Dosya başarıyla silindi. Key: %s\n", key)
+	return nil
 }
 
 func (s *R2Service) GetPresignedURL(key string, expires time.Duration) (string, error) {
@@ -79,4 +99,30 @@ func (s *R2Service) GetPresignedURL(key string, expires time.Duration) (string, 
 	}
 
 	return request.URL, nil
+}
+
+// GetFile dosyanın içeriğini byte array olarak döndürür
+func (s *R2Service) GetFile(key string) ([]byte, error) {
+	fmt.Printf("R2 - Dosya okuma işlemi başlatıldı. Key: %s\n", key)
+
+	result, err := s.client.GetObject(context.TODO(), &s3.GetObjectInput{
+		Bucket: aws.String(s.bucketName),
+		Key:    aws.String(key),
+	})
+
+	if err != nil {
+		fmt.Printf("R2 - HATA: Dosya okunurken hata oluştu. Key: %s, Hata: %v\n", key, err)
+		return nil, err
+	}
+	defer result.Body.Close()
+
+	// Dosya içeriğini oku
+	body, err := io.ReadAll(result.Body)
+	if err != nil {
+		fmt.Printf("R2 - HATA: Dosya içeriği okunurken hata oluştu. Key: %s, Hata: %v\n", key, err)
+		return nil, err
+	}
+
+	fmt.Printf("R2 - Dosya başarıyla okundu. Key: %s, Boyut: %d bytes\n", key, len(body))
+	return body, nil
 }
